+++ src/pages/CompletePage/CompletePage.module.css
@@ -1,0 +1,5 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 25px;
+}
