+++ src/pages/HomePage/HomePage.module.css
@@ -1,0 +1,50 @@
+.wrapper {
+  display: flex;
+  flex-direction: column;
+
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  padding: 40px;
+  width: 370px;
+}
+.app {
+  height: 700px;
+  display: flex;
+  justify-content: center;
+}
+.inputSectionWrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+  overflow: scroll;
+}
+
+::-webkit-scrollbar {
+  display: none;
+}
+
+.inputWrapper {
+  display: flex;
+  gap: 10px;
+}
+
+.inputTitle {
+  font-size: 12px;
+}
+
+.errorMessage {
+  font-size: 9.5px;
+  margin-top: 8px;
+  color: var(--color-red);
+}
+
+.titleWrapper h1 {
+  font-weight: 700;
+  font-size: 18px;
+}
+
+.titleWrapper p {
+  font-weight: 400;
+  font-size: 9.5px;
+  color: var(--color-gray-20);
+  margin-bottom: 16px;
+}
